requirements.txt
Pillow==6.2.0
opencv-python==4.1.2.30
numpy==1.17.4